+++ NEWS.md
@@ -5,6 +5,7 @@
 - fix trend order of area plot both global and country tabs, added utils funtion (#74)
 - correct plot titles for modules mod_compare_nth_cases_plot and mod_growth_death_rate (#66)
 - Add angle to bar-plot labels (#68)
+- Add status split stackedbarplot (#83)
 
 ### Covid19 1.1.6 (2020-04-06)